services/user_service/requirements.txt
# User Service Requirements
Flask>=2.2.0
Flask-SQLAlchemy>=3.0.0
pymysql>=1.0.2
python-dotenv>=1.0.0